requirements/base.txt
Django==5.2.7
django-environ==0.12.0
django-role-permissions==3.2.0
django-multiselectfield==1.0.1

whitenoise==6.11.0

celery==5.5.3

django-filter==25.2
django-extensions==4.1
django-cors-headers==4.9.0
loguru==0.7.3

attrs==25.4.0

gunicorn==23.0.0

requests==2.32.5

pillow==12.0.0